requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
dataclasses==0.8
Deprecated==1.2.12
fastapi==0.65.2
filelock==3.0.12
gitdb==4.0.7
GitPython==3.1.17
h11==0.12.0
huggingface-hub==0.0.8
idna==2.10
importlib-metadata==4.5.0
install==1.3.4
joblib==1.0.1
numpy==1.19.5
packaging==20.9
pycparser==2.20
pydantic==1.8.2
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scipy==1.5.4
six==1.16.0
smmap==4.0.0
starlette==0.14.2
tokenizers==0.10.3
torch==1.8.1
tqdm==4.61.1
transformers==4.6.1
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
wrapt==1.12.1
zipp==3.4.1
